backend/requirements-311.txt
# AXONENGINE v4.0 - Core Dependencies
# ============================================
# OPTION A: Python 3.11 / 3.12 (RECOMMENDED)
# ============================================
# Status: FULLY TESTED, PRODUCTION READY
# Python: 3.11.x or 3.12.x
# Compatibility: 100% - all packages have stable wheels
# Installation time: ~3-5 minutes
# 
# USE THIS IF:
# - You're on Python 3.11 or 3.12 (recommended)
# - You want fastest, most stable setup
# - You don't need Python 3.14 features
#
# SETUP:
# py -3.11 -m venv .venv
# .venv\Scripts\Activate.ps1
# pip install -r requirements-3.11.txt

# ============================================
# WEB FRAMEWORK & SERVER
# ============================================
fastapi==0.115.6
uvicorn[standard]==0.32.1
sse-starlette==2.1.3
pydantic==2.10.3
python-dotenv==1.0.1

# ============================================
# LLM & REASONING (LangChain/LangGraph Stack)
# ============================================
langchain==0.3.13
langchain-openai==0.3.1
langchain-groq==0.2.3
langgraph==0.2.60
langsmith==0.1.147

# ============================================
# LLM PROVIDERS
# ============================================
openai==1.59.3
groq==0.13.0
together==1.3.11

# ============================================
# CHEMISTRY & BIOLOGY (PRODUCTION TESTED)
# ============================================
# rdkit-pypi: Pre-built wheels for 3.11/3.12 (no compilation needed)
rdkit-pypi==2024.3.6
biopython==1.84

# ============================================
# MOLECULAR DOCKING & STRUCTURE PREPARATION
# ============================================
# vina: AutoDock Vina for molecular docking
vina==1.2.5

# meeko: Removed - has build issues on Windows
# Use standalone meeko installation if needed:
#   conda install -c conda-forge meeko

# ============================================
# DATABASES & ASYNC
# ============================================
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.36

# ============================================
# HTTP REQUESTS
# ============================================
requests==2.32.3
httpx==0.28.1

# ============================================
# NUMERICAL & IMAGE PROCESSING
# ============================================
# numpy: Updated constraint - enables NumPy 2.x which is better optimized
numpy>=1.22.4
pillow>=10.3.0,<11.0.0
reportlab==4.2.5

# ============================================
# SYSTEM DEPENDENCIES (install separately)
# ============================================
# openbabel - for SMILES/PDB/PDBQT conversion
#   Windows: choco install openbabel OR download MSI
#   Ubuntu/Debian: apt install openbabel
#   macOS: brew install open-babel
#   Verify: obabel -V
#
# fpocket - for pocket detection (optional)
#   Ubuntu/Debian: apt install fpocket
#   Requires: netcdf from conda-forge
#   Note: Not available on Windows without conda